Common base for activator
