+++ src/client/activation/common/downloadChannelRules.ts
