+++ src/test/activation/languageServer/downloadChannelRules.unit.test.ts
@@ -7,7 +7,7 @@ import { expect } from 'chai';
 import * as path from 'path';
 import { SemVer } from 'semver';
 import * as typeMoq from 'typemoq';
-import { DownloadBetaChannelRule, DownloadDailyChannelRule, DownloadStableChannelRule } from '../../../client/activation/languageServer/downloadChannelRules';
+import { DownloadBetaChannelRule, DownloadDailyChannelRule, DownloadStableChannelRule } from '../../../client/activation/common/downloadChannelRules';
 import { IPersistentState, IPersistentStateFactory } from '../../../client/common/types';
 import { IServiceContainer } from '../../../client/ioc/types';
